refactor: SectionsTest -> SectionTest로 이동
